feat: Add game over state and UI
Introduce a new `GameOver` state to the game flow. This state is displayed when the maximum number of questions has been reached.

The `GameOver` component is added to provide a celebratory message and allow players to start a new game. The `GameBoard` now displays the current question number out of the total. The `GameState` enum and URL parsing logic have been updated to accommodate the new state. A constant `MAX_QUESTIONS` is introduced.